# tools/null-g-proxy/openapi.yaml
openapi: 3.0.3
info:
  title: Antigravity Proxy API
  description: |
    OpenAI-compatible HTTP proxy for the Antigravity IDE.
    Exposes chat completions, Git helpers, a knowledge base, terminal execution,
    and code intelligence — all from a single local REST API.
  version: 1.0.0
  contact:
    name: null-g-proxy
  license:
    name: MIT
servers:
  - url: http://localhost:8787
    description: Local development server
...
paths:
...
  /v1/code/search:
    get:
      tags: [Code]
      summary: Search codebase
      description: |
        Searches the codebase using **ripgrep** (if available) or falls back to **grep**.
        Returns file path, line number, column, and matched text for each result.
      operationId: searchCode
      parameters:
        - name: q
          in: query
          required: true
          description: Search query / pattern
          schema:
            type: string
          example: resolveInstance
        - name: workspace
          in: query
          required: false
          description: Directory to search (defaults to `$HOME`)
          schema:
            type: string
          example: /Users/user/workspace/null-g-proxy
        - name: max
          in: query
          required: false
          description: Maximum number of results (default 50)
          schema:
            type: integer
            default: 50
          example: 20
        - name: glob
          in: query
          required: false
          description: File glob filter (e.g. `*.ts`)
          schema:
            type: string
          example: "*.ts"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeSearchResponse'
              example:
                query: resolveInstance
                workspace: /Users/user/workspace/null-g-proxy
                results:
                  - file: /Users/user/workspace/null-g-proxy/src/index.ts
                    line: 22
                    column: 10
                    text: "import { resolveInstance, type AntigravityInstance } from './discovery.js';"
                total: 1
                engine: ripgrep
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/code/lint:
    get:
      tags: [Code]
      summary: Lint a file
      description: |
        Lints the given file using **ESLint** (preferred) or **TypeScript compiler** (`tsc --noEmit`).
        
        The project root is auto-detected by walking up the directory tree to find `package.json`.
        Returns an empty issues array if no linter is found.
      operationId: lintFile
      parameters:
        - name: file
          in: query
          required: true
          description: Absolute path to the file to lint
          schema:
            type: string
          example: /Users/user/workspace/null-g-proxy/src/index.ts
      responses:
        '200':
          description: Lint results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LintResponse'
              example:
                file: /Users/user/workspace/null-g-proxy/src/index.ts
                issues:
                  - file: /Users/user/workspace/null-g-proxy/src/index.ts
                    line: 42
                    column: 5
                    severity: warning
                    message: "'x' is defined but never used."
                    rule: "@typescript-eslint/no-unused-vars"
                summary:
                  errors: 0
                  warnings: 1
                  total: 1
                linter: eslint
        '400':
          $ref: '#/components/responses/BadRequest'

# ─── Components ──────────────────────────────────────────────────────────────
components:
  schemas:
...
    SearchMatch:
      type: object
      properties:
        file:
          type: string
        line:
          type: integer
        column:
          type: integer
        text:
          type: string
          description: Matched line text
    CodeSearchResponse:
      type: object
      properties:
        query:
          type: string
        workspace:
          type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchMatch'
        total:
          type: integer
        engine:
          type: string
          enum: [ripgrep, grep]
    LintIssue:
      type: object
      properties:
        file:
          type: string
        line:
          type: integer
        column:
          type: integer
        severity:
          type: string
          enum: [error, warning, info]
        message:
          type: string
        rule:
          type: string
          nullable: true
    LintResponse:
      type: object
      properties:
        file:
          type: string
        issues:
          type: array
          items:
            $ref: '#/components/schemas/LintIssue'
        summary:
          type: object
          properties:
            errors:
              type: integer
            warnings:
              type: integer
            total:
              type: integer
        linter:
          type: string
          enum: [eslint, tsc, none]

    # Errors
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            type:
              type: string
            code:
              type: string
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              message: "`q` query parameter is required"
              type: invalid_request_error
...